Split up client block placement packet into multiple utility classes
